Add files via upload
In programming, loops are used to execute a block of code repeatedly based on a certain condition or for a specific number of times. Python provides two types of loops, for and while loops, that allow you to repeat code multiple times.
For Loops:-
A for loop is used to iterate over a sequence, such as a list, tuple, or string.
Range Function:-
The range() function is commonly used with for loops in Python. The range() function returns a sequence of numbers, starting from 0 by default, and increments by 1 by default, and ending at a specified number.
While Loops:-
A while loop is used to execute a block of code repeatedly as long as a certain condition is true.